support minification
After the minification Angular cannot deduce the names of dependencies from the names of parameters (they are minified). Angular also offers an official way to work around this. It can be applied automatically with ng-annotate, but this requires an extra dependency and build slowness. 

I propose to include names of dependencies in the source, as per Angular documentation.